requirements.txt
python-dotenv==1.0.0
python-telegram-bot==20.2
detectlanguage~=1.5.0